Extracted BaseSQlInput
